backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Werkzeug>=3.0.1
spacy>=3.7.2
python-docx>=0.8.11
pdfplumber>=0.10.3
PyPDF2>=3.0.1
pymupdf>=1.23.0
phonenumbers>=8.13.0
scikit-learn>=1.3.2
joblib>=1.4.2
numpy>=1.26.0
sentence-transformers>=2.2.2
regex>=2023.10.3
flask-smorest>=0.42.0
flask-limiter>=3.5.0
redis>=5.0.0
marshmallow>=3.20.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
gunicorn>=21.2.0;sys_platform!="win32"

# ── Phase 10: Production Optimization ──
celery[redis]>=5.3.0
Flask-SocketIO>=5.3.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
optimum[onnxruntime]>=1.16.0
onnxruntime>=1.17.0
prometheus-client>=0.20.0
eventlet>=0.35.0
